Refactor file/package structure
