+++ pkg/lint/rulesmeta/rules/testdata/8_missing_titles.json
