+++ pkg/lint/rulesmeta/rules/testdata/has_titles.json
